refactor(fieldmap): extract plot tile rendering into PlotTile component
